Move git backend legacy packages.
This commit performs some codebase cleanup. The `mdstreams`,
`decredplugin`, and `cmsplugin` packages are all legacy git backend
packages that will be fully removed once CMS migrates off of the git
backend. Until then, these packages have been moved out of the main
politeia directory and in with the rest of the legacy politeia code.
