+++ politeiad/backend/gitbe/decredplugin/decredplugin.go
