+++ politeiad/backend/gitbe/decredplugin/decredplugin_test.go
