+++ politeiad/backend/gitbe/gitbe.go
@@ -26,12 +26,12 @@ import (
 	"github.com/decred/dcrd/chaincfg/v3"
 	v1 "github.com/decred/dcrtime/api/v1"
 	"github.com/decred/dcrtime/merkle"
-	"github.com/decred/politeia/cmsplugin"
-	"github.com/decred/politeia/decredplugin"
 	pd "github.com/decred/politeia/politeiad/api/v1"
 	"github.com/decred/politeia/politeiad/api/v1/identity"
 	"github.com/decred/politeia/politeiad/api/v1/mime"
 	"github.com/decred/politeia/politeiad/backend"
+	"github.com/decred/politeia/politeiad/backend/gitbe/cmsplugin"
+	"github.com/decred/politeia/politeiad/backend/gitbe/decredplugin"
 	"github.com/decred/politeia/util"
 	filesystem "github.com/otiai10/copy"
 	"github.com/robfig/cron"
