+++ politeiawww/legacy/invoices.go
@@ -22,9 +22,9 @@ import (
 
 	"github.com/decred/dcrd/dcrutil/v3"
 	"github.com/decred/dcrtime/merkle"
-	"github.com/decred/politeia/decredplugin"
 	pd "github.com/decred/politeia/politeiad/api/v1"
 	"github.com/decred/politeia/politeiad/api/v1/identity"
+	"github.com/decred/politeia/politeiad/backend/gitbe/decredplugin"
 	cms "github.com/decred/politeia/politeiawww/api/cms/v1"
 	www "github.com/decred/politeia/politeiawww/api/www/v1"
 	"github.com/decred/politeia/politeiawww/legacy/cmsdatabase"
